refactor: migrate GamesPage, LeaderboardPage, HomePage to TanStack Query
- GamesPage: Replace manual fetch/useEffect with useQuery + gamesQueryOptions
  - Add pagination support with keepPreviousData for smooth transitions
  - Add error handling with retry button
  - Remove ~40 lines of boilerplate state management

- LeaderboardPage: Replace manual fetch with useQuery + leaderboardQueryOptions
  - Add error handling
  - Remove local type definitions (now imported from queries/)

- HomePage: Replace stats useEffect with useQuery + homeStatsQueryOptions
  - Stats now cached and shared across components
  - Remove local HomeStats interface (now imported from queries/)

Benefits:
- Automatic caching (navigate back without refetching)
- Background refetching when data becomes stale
- Built-in error retries
- DevTools for debugging
- ~100 lines of boilerplate code removed
