+++ client/src/components/HomePage.tsx
@@ -1,11 +1,13 @@
 import { lazy, Suspense, useEffect, useRef, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
+import { useQuery } from '@tanstack/react-query';
 import {
   loadBotGameRoute,
   loadLocalGameRoute,
   loadQuickPlayRoute,
 } from '../lib/routePrefetch';
 import { liveGameRoute, routes } from '../lib/routes';
+import { homeStatsQueryOptions, type HomeStats } from '../queries/stats';
 
 import { useTranslation } from '../lib/i18n';
 import { usePublicLiveGames } from '../hooks/usePublicLiveGames';
@@ -46,9 +48,6 @@ const SHOWCASE_PIECES: { type: PieceType; color: PieceColor }[] = [
   { type: 'P', color: 'white' },
 ];
 
-interface HomeStats {
-  totalGames: number;
-}
 type SocketModule = typeof import('../lib/socket');
 type SocketLike = SocketModule['socket'];
 
@@ -66,8 +65,14 @@ export default function HomePage() {
   const [deferredContentReady, setDeferredContentReady] = useState(import.meta.env.MODE === 'test');
   const [showDeferredContent, setShowDeferredContent] = useState(false);
   const [showHeroDecor, setShowHeroDecor] = useState(import.meta.env.MODE === 'test');
-  const [stats, setStats] = useState<HomeStats | null>(null);
   const { games: liveGames, loading: liveGamesLoading } = usePublicLiveGames({ status: 'live', limit: 4, enabled: showDeferredContent });
+  
+  // Use TanStack Query for stats
+  const { data: stats } = useQuery({
+    ...homeStatsQueryOptions(),
+    enabled: showDeferredContent,
+  });
+  
   const gameCreatedHandlerRef = useRef<((payload: { gameId: string }) => void) | null>(null);
   const connectHandlerRef = useRef<(() => void) | null>(null);
   const errorHandlerRef = useRef<((payload: { message: string }) => void) | null>(null);
@@ -98,19 +103,6 @@ export default function HomePage() {
       cleanupCreateHandlers();
     };
   }, []);
-
-  useEffect(() => {
-    if (!showDeferredContent || typeof fetch !== 'function') return;
-
-    fetch('/api/stats')
-      .then((response) => response.json())
-      .then((data) => {
-        if (typeof data?.totalGames === 'number') {
-          setStats({ totalGames: data.totalGames });
-        }
-      })
-      .catch(() => {});
-  }, [showDeferredContent]);
 
   useEffect(() => {
     if (deferredContentReady || typeof window === 'undefined') return;
